--- Common/GlobalPlayer/BlackjackPlayer.cs
@@ -1,13 +1,0 @@
-﻿using Terraria.ModLoader;
-
-namespace Blackjack.Common.GlobalPlayer
-{
-    internal class BlackjackPlayer : ModPlayer
-    {
-        public override void OnEnterWorld()
-        {
-            base.OnEnterWorld();
-
-        }
-    }
-}
